Refactor runtime bootstrap and messaging orchestration
